refactor: Consolidate expression and common rendering logic
